+++ egui/src/containers/collapsing_header.rs
@@ -133,6 +133,7 @@ pub struct CollapsingHeader {
 }
 
 impl CollapsingHeader {
+    /// The `CollapsingHeader` starts out collapsed unless you call `default_open`.
     pub fn new(label: impl Into<String>) -> Self {
         let label = Label::new(label)
             .text_style(TextStyle::Button)