+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added ⭐
+
+* `SelectableLabel` (`ui.selectable_label` and `ui.selectable_value`): a text-button that can be selected
+
 ## 0.4.0 - 2020-11-28
 
 ### Added ⭐